feat: 초기 카드를 전달하는 기능 추가
