+++ README.md
@@ -15,9 +15,9 @@
 ### 카드 뽑기
 - [x] 게임에 참여할 사람의 목록을 입력 받으면 모든 종류의 카드를 생성한다.
 - [x] 생성한 카드를 랜덤으로 섞는다.
+- [x] 카드를 나누어줄 때에는 랜덤으로 섞은 카드를 뒤에서부터 하나씩 준다.
 - [ ] 딜러와 참가자들에게 각 2장의 카드를 나눠준다.
-- [ ] 카드를 나누어줄 때에는 랜덤으로 섞은 카드를 뒤에서부터 하나씩 준다.
-- [ ] 참가자들 순서대로 카드를 더 뽑을지 확인한다.
+- [ ] 참가자들 순서대로 카드를 더 g뽑을지 확인한다.
 - [ ] 현재 참가자의 숫자 총 합이 21 아래이며 카드 여분이 있을 경우에 뽑을지 확인한다.
 - [ ] `y` 를 입력할 경우 카드를 하나 추가한다.
 - [ ] `n` 을 입력할 경우 다음 참가자로 넘어간다.
