+++ src/test/java/domain/CardsTest.java
@@ -5,12 +5,12 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CardMachineTest {
+public class CardsTest {
 
     @DisplayName("52개의 생성된 카드 개수가 올바른지 확인한다.")
     @Test
     void select_cards() {
-        CardMachine cardMachine = new CardMachine();
+        Cards cardMachine = new Cards();
         cardMachine.generate();
 
         assertThat(cardMachine.getCards().size()).isEqualTo(52);
